pyproject.toml
[project]
name = "llm_harness"
version = "0.1.0"
description = "Personal LLM harness"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    "docling>=2.82.0",
    "exa-py>=2.10.2",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-openrouter>=0.2.0",
    "langgraph>=1.1.3",
    "numpy>=2.4.4",
    "opencc-python-reimplemented>=0.1.7",
    "polars>=1.39.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "ruff>=0.15.8",
    "scipy>=1.17.1",
]

[project.optional-dependencies]
image = [
    "pillow>=12.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llm_harness"]

[tool.ruff]
line-length = 180

[tool.ruff.lint]
select = [
    "E4", "E7", "E9", "F",   # Pyflakes, pycodestyle
    "B", "B9",               # flake8-bugbear
    "I",                     # isort
    "UP",                    # pyupgrade
    "SIM",                   # flake8-simplify
    "C4",                    # flake8-comprehensions
    "ARG",                   # flake8-unused-arguments
    "DTZ",                   # flake8-datetimez
    "G",                     # flake8-logging-format
    "S",                     # flake8-bandit
    "RUF",                   # Ruff-specific rules
    "PT",                    # flake8-pytest-style
    "PIE",                   # flake8-pie
    "T20",                   # flake8-print
]
ignore = [
    "B008", "S101", "S105", "S106", "G004", "T201", "S112", "E731", "S311", "I001", "S608"
]
unfixable = ["DTZ", "S"]

[tool.ruff.lint.isort]
known-first-party = ["package_name"]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
docstring-code-format = true
